Refactor ContactService, SearchContactComponent, and ContactComponent (#6)
* feat: Add HttpClientModule import to enable HTTP requests in HomeComponent

* feat: Persist selected contact in localStorage across page reloads

* fix: Optimize contact data management with contact$ observable

Refactor the contact management to utilize a contact$ observable in the ContactService. This change allows for more efficient and responsive updates when changes occur in the backend database (db.js)

* fix: Centralize ContactService provider in AppComponent for shared state management
- Removed individual ContactService providers from each component to address an issue where updates to contacts$ observable were not being detected across components.
- Set ContactService as a provider in AppComponent to ensure a single, application-wide instance, allowing all components to share and reflect a consistent state.

* feat: Add tap and map operators to update contacts$ observable with new data
- Implement tap and map operators in ContactService to transform and update the contacts$ observable.
- This change ensures that any updates to the contact data are reflected in the observable, improving data consistency across the application.

* feat: Add tap and filter operators in deleteContact method to update contacts$ observable with new data
- Implement tap and filter operators in deleteContact method to transform and update the contacts$ observable.
- This change ensures that any updates to the contact data are reflected in the observable, improving data consistency across the application.

* fear: Manage selected contact Id using BehaviorSubject instead of localStorage

* refactor: Simplify ContactComponent by removing redundant error handling
- Removed unnecessary catchError in ngOnInit.
- Leveraged error handling from service methods.
- Consolidated subscriptions for easier management and cleanup.

* refactor: Add error handling in ContactService
- Implemented catchError in all HTTP request methods.
- Added handleError method to log errors and provide default values.

* feat: Refactor SearchContactComponent to manage state with BehaviorSubject
- Updated to use BehaviorSubject for managing selected contact ID.
- Added setupSearchSubscription, setupContactsSubscription, and setupSelectedContactSubscription methods for better readability.
- Subscribed to selectedContactId$ observable.
- Removed localStorage usage for storing the selected contact ID.
- Ensured proper cleanup of subscriptions in ngOnDestroy.

+++ src/app/home/ui/components/contact/contact.component.ts
@@ -1,21 +1,22 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Contact } from '../../../../shared/interface/contact.interface';
 import { ContactService } from '../../../data-access/contact.service';
-import { catchError, switchMap } from 'rxjs/operators';
+import { switchMap } from 'rxjs/operators';
 import { CommonModule } from '@angular/common';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-contact',
   standalone: true,
-  imports: [CommonModule],
+  imports:[CommonModule],
   templateUrl: './contact.component.html',
   styleUrl: './contact.component.css',
-  providers: [ContactService],
 })
-export class ContactComponent implements OnInit {
+export class ContactComponent implements OnInit, OnDestroy {
   userId: string = '';
   contact: Contact | undefined;
+  private subscription = new Subscription();
 
   constructor(
     private route: ActivatedRoute,
@@ -24,50 +25,54 @@ export class ContactComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.route.paramMap
-      .pipe(
-        switchMap((params) => {
-          this.userId = params.get('id') || '';
-          return this.contactService.getContactById(this.userId);
-        }),
-      )
-      .subscribe((contact) => { this.contact = contact });
+    this.subscription.add(
+      this.route.paramMap
+        .pipe(
+          switchMap((params) => {
+            this.userId = params.get('id') || '';
+            this.contactService.selectContact(this.userId)
+            return this.contactService.getContactById(this.userId);
+          })
+        ).subscribe((contact) => { this.contact = contact })
+    );
   }
 
-  handleFavorite() {
+  handleFavorite():void {
     if (this.contact) {
       const newFavoriteStatus = !this.contact.favorite;
-      this.contactService
-        .updateContact(this.userId, { favorite: newFavoriteStatus })
-        .subscribe({
-          next: (updatedContact) => {
-            this.contact = updatedContact;
-            this.contact.favorite = newFavoriteStatus;
-          },
-          error: (err) => {
-            console.error('Error:', err);
-          },
-        });
+      this.subscription.add(
+        this.contactService
+          .updateContact(this.userId, { favorite: newFavoriteStatus })
+          .subscribe({
+            next: (updatedContact) => {
+              this.contact = updatedContact;
+              this.contact.favorite = newFavoriteStatus;
+            },
+            error: (err) => {
+              console.error('Error updating contact:', err);
+            },
+          })
+      )
     }
   }
 
   handleDelete() {
-    if (this.contact) {
-      if (
-        confirm(
-          `Are you sure you want to delete ${this.contact.name} ${this.contact.lastname}?`
-        )
-      ) {
+    if (this.contact && confirm(`Are you sure you want to delete ${this.contact.name} ${this.contact.lastname}?`)) {
+      this.subscription.add(
         this.contactService.deleteContact(this.userId).subscribe({
           next: (contact) => {
             alert(`${contact.name} ${contact.lastname} has been deleted.`);
             this.router.navigate(['']);
           },
           error: (err) => {
-            console.error('Error:', err);
+            console.error('Error deleting contact:', err);
           },
-        });
-      }
+        })
+      )
     }
+  }
+
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe()
   }
 }
